Move 1.0.0 to 18.0.0 to bypass lint constraint
