+++ sdk/go/pulumi-language-go/testdata/sdks/union-18.0.0/go.mod
@@ -1,4 +1,4 @@
-module example.com/pulumi-union/sdk/go
+module example.com/pulumi-union/sdk/go/v18
 
 go 1.20
