+++ sdk/go/pulumi-language-go/testdata/sdks/union-18.0.0/union/internal/pulumiVersion.go
