+++ sdk/go/pulumi-language-go/testdata/sdks/union-18.0.0/union/provider.go
@@ -7,7 +7,7 @@ import (
 	"context"
 	"reflect"
 
-	"example.com/pulumi-union/sdk/go/union/internal"
+	"example.com/pulumi-union/sdk/go/v18/union/internal"
 	"github.com/pulumi/pulumi/sdk/v3/go/pulumi"
 )
