+++ sdk/nodejs/cmd/pulumi-language-nodejs/testdata/tsc/sdks/union-18.0.0/index.ts
