+++ sdk/nodejs/cmd/pulumi-language-nodejs/testdata/tsc/sdks/union-18.0.0/package.json
@@ -1,6 +1,6 @@
 {
     "name": "@pulumi/union",
-    "version": "1.0.0",
+    "version": "18.0.0",
     "scripts": {
         "build": "tsc"
     },
@@ -14,6 +14,6 @@
     "pulumi": {
         "resource": true,
         "name": "union",
-        "version": "1.0.0"
+        "version": "18.0.0"
     }
 }
