+++ sdk/nodejs/cmd/pulumi-language-nodejs/testdata/tsnode/sdks/union-18.0.0/provider.ts
