+++ sdk/python/cmd/pulumi-language-python/testdata/classes/projects/l2-union/requirements.txt
@@ -1,3 +1,3 @@
 ROOT/artifacts/pulumi-CORE.VERSION-py3-none-any.whl
-ROOT/artifacts/pulumi_union-1.0.0-py3-none-any.whl
+ROOT/artifacts/pulumi_union-18.0.0-py3-none-any.whl
 pyright
