+++ sdk/python/cmd/pulumi-language-python/testdata/classes/sdks/union-18.0.0/pulumi_union/__init__.py
