+++ sdk/python/cmd/pulumi-language-python/testdata/setuppy/sdks/union-18.0.0/pulumi_union/example.py
