+++ sdk/python/cmd/pulumi-language-python/testdata/setuppy/sdks/union-18.0.0/pulumi_union/pulumi-plugin.json
@@ -1,5 +1,5 @@
 {
   "resource": true,
   "name": "union",
-  "version": "1.0.0"
+  "version": "18.0.0"
 }
