+++ sdk/python/cmd/pulumi-language-python/testdata/setuppy/sdks/union-18.0.0/setup.py
@@ -9,7 +9,7 @@ from setuptools.command.install import install
 from subprocess import check_call
 
 
-VERSION = "1.0.0"
+VERSION = "18.0.0"
 def readme():
     try:
         with open('README.md', encoding='utf-8') as f:
