+++ sdk/python/cmd/pulumi-language-python/testdata/toml/sdks/union-18.0.0/pulumi_union/_utilities.py
@@ -328,4 +328,4 @@ def get_plugin_download_url():
 	return None
 
 def get_version():
-    return "1.0.0"
+    return "18.0.0"
