+++ sdk/python/cmd/pulumi-language-python/testdata/toml/sdks/union-18.0.0/pulumi_union/provider.py
