+++ sdk/python/cmd/pulumi-language-python/testdata/toml/sdks/union-18.0.0/pyproject.toml
@@ -3,7 +3,7 @@
   dependencies = ["parver>=0.2.1", "pulumi>=3.165.0,<4.0.0", "semver>=2.8.1", "typing-extensions>=4.11,<5; python_version < \"3.11\""]
   readme = "README.md"
   requires-python = ">=3.9"
-  version = "1.0.0"
+  version = "18.0.0"
 
 [build-system]
   requires = ["setuptools>=61.0"]
